+++ neoforge/src/main/java/fr/loudo/narrativecraft/events/client/OnClientTickEventNeoForge.java
@@ -25,12 +25,13 @@ package fr.loudo.narrativecraft.events.client;
 
 import fr.loudo.narrativecraft.NarrativeCraftMod;
 import net.minecraft.client.Minecraft;
+import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.neoforge.client.event.ClientTickEvent;
 import net.neoforged.neoforge.common.NeoForge;
 
-@Mod(NarrativeCraftMod.MOD_ID)
+@Mod(value = NarrativeCraftMod.MOD_ID, dist = Dist.CLIENT)
 public class OnClientTickEventNeoForge {
 
     public OnClientTickEventNeoForge(IEventBus eventBus) {