+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - Can't drop item with NeoForge
+- Every entity on a playback had a visible name (like items dropped)
 
 ## [2.0.11]