+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 # Changelog
 
-## [2.0.14]
+## [2.0.15]
 
 ### Fixes
 
-- Can't drop item with NeoForge
-- Every entity on a playback had a visible name (like items dropped)
+- NeoForge server not starting